requirements.txt
# Core scientific computing
numpy>=1.24,<3
pandas>=2.0,<3
scipy>=1.10,<2

# Machine learning
scikit-learn>=1.3,<2

# YAML parsing
PyYAML>=6.0,<7

# Testing
pytest>=7.4,<9
pytest-cov>=4.0,<6

# Linting
flake8>=6.0,<8

# Visualization (optional but used in examples)
matplotlib>=3.7,<4

# PyTorch is optional — install separately for layout model training
# pip install torch  (CPU-only) or see https://pytorch.org/get-started/locally/
